Keep MCP resource registration isolated from server tools
Move the resource decorators into a dedicated module and re-export the functions from mcp_video.server so existing imports continue to resolve. The server module now imports registered resources once instead of owning duplicate decorators.

Constraint: Preserve from mcp_video.server resource imports while avoiding duplicate MCP registration

Rejected: Split tool handlers too | outside the requested small server resource split

Confidence: high

Scope-risk: narrow

Tested: python3 -m pytest tests/test_server.py tests/test_public_surface.py -q (92 passed)

Tested: python3 -m ruff check mcp_video tests

Tested: python3 -c "import mcp_video"

Not-tested: python3 -m pytest tests/ -x -q --tb=short blocked by missing out/McpVideoExplainer-FINAL.mp4 fixture after 538 passed and 9 skipped
