+++ mcp_video/server.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import json
 import os
 import re
 from typing import Any
@@ -47,6 +46,12 @@ from .engine import (
 )
 from .errors import MCPVideoError
 from .server_app import _error_result, _result, mcp
+from .server_resources import (
+    templates_resource as templates_resource,
+    video_audio_resource as video_audio_resource,
+    video_info_resource as video_info_resource,
+    video_preview_resource as video_preview_resource,
+)
 from .limits import (
     MAX_BATCH_SIZE,
     MAX_CONCURRENCY,
@@ -79,82 +84,6 @@ from .validation import (
     VALID_WAVEFORMS,
     VALID_WHISPER_MODELS,
 )
-
-
-# ---------------------------------------------------------------------------
-# MCP Resources
-# ---------------------------------------------------------------------------
-
-
-@mcp.resource("mcp-video://video/{path}/info")
-def video_info_resource(path: str) -> str:
-    """Get metadata about a video file (duration, resolution, codec, etc.)."""
-    try:
-        info = probe(path)
-        return info.model_dump_json(indent=2)
-    except MCPVideoError as e:
-        return json.dumps(_error_result(e), indent=2)
-
-
-@mcp.resource("mcp-video://video/{path}/preview")
-def video_preview_resource(path: str) -> str:
-    """Get a text storyboard description (key frame timestamps)."""
-    try:
-        info = probe(path)
-        frames = []
-        dur = info.duration
-        count = 8
-        for i in range(count):
-            ts = dur * (i + 1) / (count + 1)
-            frames.append(f"Frame {i + 1}: {ts:.1f}s")
-        return "\n".join(frames)
-    except MCPVideoError as e:
-        return json.dumps(_error_result(e), indent=2)
-
-
-@mcp.resource("mcp-video://video/{path}/audio")
-def video_audio_resource(path: str) -> str:
-    """Extract and describe the audio track of a video."""
-    try:
-        info = probe(path)
-        if info.audio_codec:
-            return (
-                f"Audio codec: {info.audio_codec}\n"
-                f"Sample rate: {info.audio_sample_rate} Hz\n"
-                f"Duration: {info.duration:.1f}s"
-            )
-        return "No audio track found."
-    except MCPVideoError as e:
-        return json.dumps(_error_result(e), indent=2)
-
-
-@mcp.resource("mcp-video://templates")
-def templates_resource() -> str:
-    """List available editing templates (aspect ratios, quality presets)."""
-    from .models import ASPECT_RATIOS, QUALITY_PRESETS
-    import json
-
-    data = {
-        "aspect_ratios": {k: f"{v[0]}x{v[1]}" for k, v in ASPECT_RATIOS.items()},
-        "quality_presets": {
-            k: f"CRF {v['crf']}, preset={v['preset']}, max_height={v['max_height']}" for k, v in QUALITY_PRESETS.items()
-        },
-        "transition_types": ["fade", "dissolve", "wipe-left", "wipe-right", "wipe-up", "wipe-down"],
-        "export_formats": ["mp4", "webm", "gif", "mov"],
-        "text_positions": [
-            "top-left",
-            "top-center",
-            "top-right",
-            "center-left",
-            "center",
-            "center-right",
-            "bottom-left",
-            "bottom-center",
-            "bottom-right",
-        ],
-    }
-    return json.dumps(data, indent=2)
-
 
 # ---------------------------------------------------------------------------
 # MCP Tools
